Add Redux Toolkit and Slicing (#24)
## Related to
- Go from redux to redux toolkit
- Adding slicing functions (add, delete slice, and give a name to the slice)

## Description
- Changing the reducer, deleting the redux actions and rewrite it with redux toolkit implementations
- Add, delete, and name a slice

## Tests
For the slicing, you need to click on the pizza slice on the tree of attributes (it will appear on hover). A modal window will open (Dialog component), and you can choose the slice name. Then, click on submit and it will create the slice (and its child if it's a backbone element).
If you want to remove it, you click on the red cross on the tree of attributes. A modal window will open, to confirm that you want to remove the slice. On submit, the slice will be deleted.
You can test it by creating slices on deep attributes.